feat(windows): verify auto-updates with Sparkle EdDSA signatures
The Windows update path downloaded and ran installers with no signature
verification: WinSparkle 0.8.1 had no key configured, so it logged
"Using unsigned updates!" and executed whatever the appcast pointed at,
protected only by HTTPS to GitHub.

- Upgrade vendored WinSparkle 0.8.1 -> 0.9.4, which adds EdDSA
  (Ed25519) verification compatible with macOS Sparkle
- Embed the existing Sparkle public key (same pair as SUPublicEDKey in
  macos/Runner/Info.plist) as the EdDSAPub resource in Runner.rc, so
  WinSparkle rejects unsigned or tampered update downloads
- Sign the Windows installer in CI with the vendored winsparkle-tool
  and the existing SPARKLE_EDDSA_PRIVATE_KEY secret; the build fails
  if the derived public key does not match the embedded one
- Emit sparkle:edSignature/length on the Windows appcast enclosure and
  make missing signatures (macOS or Windows) a hard generator error;
  SPARKLE_ALLOW_UNSIGNED_WINDOWS=1 remains as an escape hatch for
  promoting betas built before Windows signing existed
- Record windowsEddsaSignature/windowsExeLength in beta metadata.json
  and thread them through release and promote appcast generation

Authenticode/SmartScreen signing (the install-time half of the issue)
is handled separately via SignPath Foundation.

Part of #848
