--- packages/auto_updater_windows/windows/WinSparkle-0.8.1/AUTHORS
@@ -1,12 +1,0 @@
-Maintainer:
-
-Vaclav Slavik <vaclav@slavik.io>
-
-
-Contributors:
-
-Kohan Ikin <syneryder@namesuppressed.com>
-Christian L. Jacobsen <christian@concurrency.cc>
-Littleboy <littleboy@users.sourceforge.net>
-Vasco Veloso
-Jonas Emanuel Mueller <zanoi@zanoi.net>
