+++ packages/auto_updater_windows/windows/WinSparkle-0.9.4/AUTHORS
@@ -1,0 +1,12 @@
+Maintainer:
+
+Vaclav Slavik <vaclav@slavik.io>
+
+
+Contributors:
+
+Kohan Ikin <syneryder@namesuppressed.com>
+Christian L. Jacobsen <christian@concurrency.cc>
+Littleboy <littleboy@users.sourceforge.net>
+Vasco Veloso
+Jonas Emanuel Mueller <zanoi@zanoi.net>
